requirements.txt
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
ConfigArgParse==1.7.1
Flask==3.1.1
flask-cors==6.0.0
Flask-Login==0.6.3
gevent==25.5.1
geventhttpclient==2.3.3
greenlet==3.2.2
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.37.9
locust-cloud==1.23.1
MarkupSafe==3.0.2
msgpack==1.1.0
numpy==2.2.6
pandas==2.3.0
platformdirs==4.3.8
psutil==7.0.0
pyspark==3.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
pyzmq==26.4.0
requests==2.32.3
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
